Fixes clippy warnings

+++ src/codec/parameters.rs
@@ -111,7 +111,7 @@ impl Parameters {
 
     #[inline]
     pub fn extradata(&mut self) -> &[u8] {
-        let extradata_size = unsafe { (*self.as_ptr()).extradata_size } as i32 as usize;
+        let extradata_size = unsafe { (*self.as_ptr()).extradata_size } as usize;
         unsafe { std::slice::from_raw_parts((*self.as_ptr()).extradata, extradata_size) }
     }
